refactor(issue69): split parameter_simulation_gpu_lib into optimization/gpu modules
