+++ src/optimization/gpu/__init__.py
@@ -2,3 +2,6 @@
 GPU-based optimization implementations.
 """
 
+from .parameter_simulation import find_optimal_parameters, main
+
+__all__ = ["find_optimal_parameters", "main"]
